LOTS OF BUG FIXES added JP test program changed addresses variables to integer RM's IC and VM's IC are now fully linked Changed code segment start cursor to 0x80

+++ src/machine/Pagination.java
@@ -12,7 +12,7 @@ public class Pagination {
 	}
 	
 	/* return real address */
-	public int getRealAddress(byte virtualAddress) {
+	public int getRealAddress(int virtualAddress) {
 		byte plr = RealMachineRegisters.getPLR();
 		int block = virtualAddress / 0x10;
 		int offset = virtualAddress % 0x10;
